feat(core): move auth.service to user
feat(core): remove SERVER_API_URL deps
