+++ packages/core/user/index.ts
@@ -6,3 +6,5 @@ export { AccountService } from './src/account.service';
 
 export * from './src/account.model';
 export * from './src/user.model';
+export * from './src/auth-jwt.service';
+export { AuthService } from './src/auth.service';
